docs: add flow config docs
